Start fix bugs:
When you click on a graphic (cluster) with draw tool on, the cluster does not expand anymore
Measure cursor stay on between measure
Double click on measure add a last point
Delete extent now is with a grey line instead of red
Tools panel height is set properly in FireFox
When esc key press, tools panel open or close only once (bug in Safari)
When you measure a line with 2 points, length was 0. It is now fix
When you change selection for a radio button in the legend, the layer will not show if parent checkbox is uncheck
Double click for measure in FireFox works now.